fix: Slow project root detection due to glob
We used to have `**/*.zig` as a glob pattern to detect project's root.
This is quite slow for large non-zig projects though, because the search
goes through all files to verify that the project does not contain any
`.zig` files.

Having just `build.zig` as a prject root pattern should be enough
though, because even the folder does not contain a `build.zig` file, the
individiaul files/buffers are tested against all adapters by neotest.

Resolves #14