fix: Cannot run tests when git submodule contains `build.zig`
Glob function used to return all matches of `build.zig` in a newline
separated list, which we incorrectly used as a single path.

Now glob pattern is requested to return a list of paths (by setting the
third parameter to true) and we take only the first one. This still
means you cannot run tests of submodules, only your root repo, as we
only support a single `build.zig` file for now.

Resolves #24